feat(vueify): draft datasets & projects pages
