+++ private/vue/Datasets.js
@@ -1,0 +1,7 @@
+import Vue from './Vue'
+import App from './Datasets.vue'
+
+new Vue({
+  el: '#datasets',
+  render: h => h(App),
+})
