+++ private/vue/Datasets.vue
@@ -1,0 +1,26 @@
+<template>
+  <div>
+    <h1>Datasets</h1>
+  </div>
+  
+</template>
+
+<script>
+export default {
+  name: 'Datasets',
+
+  components: {},
+  data() {
+    return Object.assign(server, {
+      maps: {},
+    })
+  },
+  created() {
+    // console.log('created')
+  },
+  methods: {},
+}
+</script>
+
+<style lang="scss" scoped>
+</style>
