--- private/vue/Layers.js
@@ -1,7 +1,0 @@
-import Vue from './Vue'
-import App from './Layers.vue'
-
-new Vue({
-  el: '#layers',
-  render: h => h(App),
-})
