--- private/vue/Layers.vue
@@ -1,26 +1,0 @@
-<template>
-  <div>
-    <h1>Layers</h1>
-  </div>
-  
-</template>
-
-<script>
-export default {
-  name: 'Layers',
-
-  components: {},
-  data() {
-    return Object.assign(server, {
-      maps: {},
-    })
-  },
-  created() {
-    // console.log('created')
-  },
-  methods: {},
-}
-</script>
-
-<style lang="scss" scoped>
-</style>
