+++ private/vue/Projects.js
@@ -1,0 +1,7 @@
+import Vue from './Vue'
+import App from './Projects.vue'
+
+new Vue({
+  el: '#projects',
+  render: h => h(App),
+})
